// src/app/api/spotify/type.ts
export type TopData = {
  topAlbum: TopAlbum;
  topTrack: TopTrack;
};

export type TopAlbum = Record<
  string,
  {
    artist: string;
    count: number;
    cover: Image;
    name: string;
  }
>;

export type TopTrack = Record<
  string,
  {
    artist: string;
    count: number;
    cover: Image;
    name: string;
  }
>;

/**
 * spotify recently-played API의 데이터 타입
 */
export interface Item {
  track: Track;
  played_at: string;
  context: Context;
}

export interface Track {
  album: Album;
  artists: TrackArtist[];
  disc_number: number;
  duration_ms: number;
  explicit: boolean;
  external_ids: ExternalIds;
  external_urls: ExternalUrls4;
  href: string;
  id: string;
  is_local: boolean;
  is_playable: boolean;
  name: string;
  popularity: number;
  preview_url: string;
  track_number: number;
  type: string;
  uri: string;
}

export interface Album {
  album_type: string;
  artists: Artist[];
  external_urls: ExternalUrls2;
  href: string;
  id: string;
  images: Image[];
  is_playable: boolean;
  name: string;
  release_date: string;
  release_date_precision: string;
  total_tracks: number;
  type: string;
  uri: string;
}

export interface Artist {
  external_urls: ExternalUrls;
  href: string;
  id: string;
  name: string;
  type: string;
  uri: string;
}

export interface ExternalUrls {
  spotify: string;
}

export interface ExternalUrls2 {
  spotify: string;
}

export interface Image {
  height: number;
  url: string;
  width: number;
}

export interface TrackArtist {
  external_urls: ExternalUrls3;
  href: string;
  id: string;
  name: string;
  type: string;
  uri: string;
}

export interface ExternalUrls3 {
  spotify: string;
}

export interface ExternalIds {
  isrc: string;
}

export interface ExternalUrls4 {
  spotify: string;
}

export interface Context {
  type: string;
  href: string;
  external_urls: ExternalUrls5;
  uri: string;
}

export interface ExternalUrls5 {
  spotify: string;
}